Get rid of build.sh, implement cache-able builds
Removes build.sh in favor of actual stages in the Dockerfile itself so
that stages can be cached instead of having to run things over and over
again on different systems.

Also adds in a patch for big endian (s390x) systems that is not
currently present in any release of protobuf.

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>